requirements.txt
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
python-telegram-bot==22.0
requests==2.32.3
service-identity==24.2.0
sniffio==1.3.1
TelegramBotAPI==0.3.2
urllib3==2.4.0
